ELI-351: De-extracts EligibilityResultBuilder and moves tests to Eligibility Calculator tests
